refactor: 성공한 문제 list를 fetch하는 로직 변경
