+++ src/pages/content/testPage.tsx
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom/client';
 import { theme } from '../../styles/theme';
 import { ThemeProvider } from 'styled-components';
 import CreateSolutionsButton from '../../components/button/CreateSolutionsButton';
-import { addSolvedProblemId } from '../background/postCurrentSolution';
 
 const $submitBtn = document.querySelector('#submit-code') as HTMLButtonElement;
 const $modal = document.querySelector('.modal') as HTMLDivElement;
@@ -71,10 +70,6 @@ const uploadCurrentSolution = async () => {
   }
 
   const data = parsingDomNodeToUpload();
-
-  const problemId = Number(data.problemId);
-  addSolvedProblemId(problemId);
-
   const uploadResult = await new Promise<boolean>(resolve => {
     chrome.runtime.sendMessage({ method: 'postCurrentSolution', data }, response => {
       resolve(response.status);
